[Flow] Add pass to fuse encoding ops into dispatch regions after hoisting
This PR is the follow up to #18063, implementing the fusion pass to move set_encoding ops into producer dispatch regions when possible. If there is no fusable producer, then the set_encoding is wrapped in a new dispatch region. The new pass reuses the fusion logic from FormDispatchRegions, treating the producer operation of the set_encoding as the dispatch root op (for the purpose of checking outer parallel loops).

Signed-off-by: hanhanW <hanhan0912@gmail.com>